+++ setup.cfg
@@ -138,7 +138,8 @@ pykeen.datasets =
     openbiolink      = pykeen.datasets.openbiolink:OpenBioLink
     openbiolinklq    = pykeen.datasets.openbiolink:OpenBioLinkLQ
     opena            = pykeen.datasets.openea:OpenEA
-    pharmkg8k        = pykeen.datasets.pharmkg8k:PharmKG8k
+    pharmkg8k        = pykeen.datasets.pharmkg:PharmKG8k
+    pharmkgfull      = pykeen.datasets.pharmkg:PharmKGFull
     umls             = pykeen.datasets.umls:UMLS
     wk3l15k          = pykeen.datasets.wk3l:WK3l15k
     wk3l120k         = pykeen.datasets.wk3l:WK3l120k
