+++ src/pykeen/datasets/__init__.py
@@ -43,7 +43,7 @@ from .nations import Nations
 from .ogb import OGBBioKG, OGBWikiKG
 from .openbiolink import OpenBioLink, OpenBioLinkLQ
 from .openea import OpenEA
-from .pharmkg8k import PharmKG8k
+from .pharmkg import PharmKG8k, PharmKGFull
 from .umls import UMLS
 from .wd50k import WD50KT
 from .wikidata5m import Wikidata5M
@@ -84,6 +84,7 @@ __all__ = [
     "WD50KT",
     "Wikidata5M",
     "PharmKG8k",
+    "PharmKGFull",
     # Utilities
     "dataset_resolver",
     "get_dataset",
